move verify account page
